Move torchhub tests into separate test_hub.py file
[ghstack-poisoned]
